イベントページに主催者アドレスを表示。NFTページの多言語対応も。
