+++ frontend/src/locales/ja.ts
@@ -84,4 +84,5 @@ export default {
   CONTRIBUTORS_WANTED_DESC:
     "MintRallyはシビックテックプロジェクトHackdaysのメンバーで開発しているオープンソースプロジェクトです。<br />機能拡充やユーザビリティの向上、アクセシビリティへの対応などに取り組みながら、よりよいオープンソースプロジェクトにしていくための仲間を探しています。",
   SHOW_NFT_ON_METAMASK: "入手したNFTをメタマスクでも見れるようにする方法",
+  OWNER: "所有者",
 };
